HID: apple: Fix Wireless Keyboards when NumLock is on
Certain Apple Wireless Keyboard models (2007, 2009, and 2011) behave
strangely when NumLock is on: The driver switches into a numeric keypad
overlay mode, locking out most keys. Worse, it is not possible to exit
this mode from the keyboard, since there is no NumLock key.

This mode is not user-friendly, since it is not obvious why key presses
translate into no input or wrong input. Users think the keyboard is
broken and have resorted to workarounds like [1] and [2].
[1] https://help.ubuntu.com/community/AppleKeyboard#Numlock_on_Apple_Wireless_Keyboard
[2] https://wiki.archlinux.org/title/Apple_Keyboard#Numlock_is_on

Follow the principle of least astonishment: These keyboards don't have
keycap labels for a numeric keypad mode, so don't implement one.
(The mode is still necessary on older keyboards with the appropriate
keycap labels for a numeric keypad mode.)

Remove the APPLE_NUMLOCK_EMULATION quirk from the 9 Wireless Keyboard
2007/2009/2011 ANSI/ISO/JIS device entries in the driver's table.

Fixes: 6e7045990f35 ("HID: split Numlock emulation quirk from HID_QUIRK_APPLE_HAS_FN.")
Fixes: 23aeb61e7e1f ("HID: add device IDs for new model of Apple Wireless Keyboard")
Fixes: ad734bc15653 ("HID: hid-apple: add device ID of another wireless aluminium")
Fixes: 0a97e1e9f9a6 ("HID: apple: Add Apple wireless keyboard 2011 ANSI PID")
Fixes: bd4a7ce1da06 ("HID: apple: add Apple wireless keyboard 2011 JIS model support")
Cc: stable@vger.kernel.org
Tested-by: Alexander F. Lent <lx@xanderlent.com>
Signed-off-by: Alexander F. Lent <lx@xanderlent.com>